bot/pyproject.toml
[project]
name = "se-toolkit-bot"
version = "0.1.0"
description = "Telegram bot for SE Toolkit Lab 7"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["handlers"]
